requirements.txt
tensorflow
google-api-python-client
oauth2client
pillow
face_recognition
click
scipy
scikit-image
six
opencv-python